refactor: move AFrameAudioVisualizer from atoms to molecules
